Generate drive client
